services/auth/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose
passlib[bcrypt]
redis
httpx
PyJWT
python-multipart
